Cast value to prevent a numeric conversion warning

+++ src/gui/painting/qdrawhelper_sse2.cpp
@@ -448,12 +448,9 @@ void qt_bitmapblit16_sse2(QRasterBuffer *rasterBuffer, int x, int y, const QRgba
    quint16 *dest = reinterpret_cast<quint16 *>(rasterBuffer->scanLine(y)) + x;
    const int destStride = rasterBuffer->bytesPerLine() / sizeof(quint16);
 
-   const __m128i c128 = _mm_set1_epi16(c);
-
-   const __m128i maskmask = _mm_set_epi16(0x0101, 0x0202, 0x0404, 0x0808,
-         0x1010, 0x2020, 0x4040, 0x8080);
-   const __m128i maskadd = _mm_set_epi16(0x7f7f, 0x7e7e, 0x7c7c, 0x7878,
-         0x7070, 0x6060, 0x4040, 0x0000);
+   const __m128i c128     = _mm_set1_epi16(c);
+   const __m128i maskmask = _mm_set_epi16(0x0101, 0x0202, 0x0404, 0x0808, 0x1010, 0x2020, 0x4040, int16_t(0x8080));
+   const __m128i maskadd  = _mm_set_epi16(0x7f7f, 0x7e7e, 0x7c7c, 0x7878, 0x7070, 0x6060, 0x4040, 0x0000);
 
    while (height--) {
       for (int x = 0; x < width; x += 8) {
